requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
apscheduler==3.10.4
aiosqlite==0.20.0
playwright==1.44.0
Pillow==10.3.0
